[opentitantool] Revert default interface to "empty"
In order to support the opentitantool subcommands that do not operate
on any debugger backend, the default backend should be one that can be
instantiated without any hardware, but does not support any features.

I previously changed that default as part of adding session support,
without realizing the usefulness of an "empty" transport.  This CL
reverts the default.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I56874bba81552cf440387d8bb110d9162946f4ce