[otbn,dv] Still print final instruction on escalation when running
We run on_lc_escalation() earlier in the cycle than step(), but we
want to make sure that we still generate a trace item for executing an
instruction on that cycle (even though it will be cancelled).

This commit adds explicit "pending error bits" to the state to make
that possible.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>