[sram_ctrl] Remove nonce reversal to improve timing
Reversing the scrambling nonces upon an integrity error has a negative
impact on timing due to the fact that the integrity error signal is one
of the signals that carries the highest combo delay among the scrambling
device input signals.

As this is not strictly needed due to transaction blocking at multiple
places (in the TL-UL adapter, in the prim_ram_1p_scr before and after
the SRAM macro), the feature is removed in order to improve timing.

Signed-off-by: Michael Schaffner <msf@opentitan.org>