[baze] invoke sim_dv logs extraction rule
This invokes the sim_dv logs extraction rule from the `opentitan_binary`
macro, and only enables emitting this rule, when enabled via a macro
flag. Additionally, this sets the flag to true when SW either ROM or
flash images are compiled for the `sim_dv` device.

This is the last task that fixes #11559.

Signed-off-by: Timothy Trippel <ttrippel@google.com>