[dif/adc_ctrl] clear filter status CSR on IRQ cause clear
As discussed, in #11354, it is more convenient to hide the ack-ing of
the filter status CSR behind the ack-ing of the IRQ status CSR.
Unfortunately, this required a HW change that was disruptive to DV at
the given moment. Therefore, SW can accomodate this by ack-ing both the
filter and IRQ status CSRs in the same DIF, which is implemented in
this commit.

Signed-off-by: Timothy Trippel <ttrippel@google.com>