[otbn] Move .start into .text section in OTBN linker script.
With .start listed as a separate section in the linker script, objcopy
was getting addresses mixed up, essentially expecting the .text.start
section to be twice as long as it actually was for some of the symbols.
Putting the `.text.start` before `.text.*` in the `.text` section
preserves the property that the start code goes first independent of
linker argument order, but fixes the symbol mixup issue.

Signed-off-by: Jade Philipoom <jadep@google.com>