[otbn, dv] Added a test to escalate software error to fatal errors
This commit adds a new testcase called otbn_sw_errs_fatal_chk. This test
sets the software_errs_fatal bit in ctrl register to escalate software
errors to fatal errors.

This commit also adds required functions to set software_errs_fatal flag
in ISS model and if this bit is set in the model, the model goes to
locked state.

Signed-off-by: Prajwala Puttappa <prajwalaputtappa@lowrisc.org>